Split up alignment view
