+++ course_flow/static/course_flow/js/react/src/Components/Views/AlignmentView/index.js
@@ -1,0 +1,1 @@
+export { default as AlignmentView } from './AlignmentView.js'
