+++ course_flow/static/course_flow/js/react/src/Components/Views/LiveProjectView/LiveProjectMenu.js
@@ -6,9 +6,9 @@ import { LiveProjectSection } from './LiveProjectSection.js'
 import { LiveProjectOverview } from './LiveProjectOverview.js'
 import LiveProjectStudents from './LiveProjectStudents.js'
 import { WorkflowForMenu } from '../../Library'
-import LiveProjectCompletionTable from "./LiveProjectCompletionTable.js";
-import LiveProjectSettings from "./LiveProjectSettings.js";
-import LiveProjectAssignments from "./LiveProjectAssignments.js";
+import LiveProjectCompletionTable from './LiveProjectCompletionTable.js'
+import LiveProjectSettings from './LiveProjectSettings.js'
+import LiveProjectAssignments from './LiveProjectAssignments.js'
 
 /**
  *
