+++ course_flow/static/course_flow/js/react/src/Components/Views/LiveProjectView/LiveProjectStudents.js
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { LiveProjectSection } from './LiveProjectSection.js'
-import StudentManagement from "../../components/StudentManagement"
+import StudentManagement from '../../components/StudentManagement'
 
 class LiveProjectStudents extends LiveProjectSection {
   render() {
