+++ course_flow/static/course_flow/js/react/src/Components/Views/OutcomeView/OutcomeView.js
@@ -172,7 +172,7 @@ class OutcomeView extends EditableComponentWithSorting {
   }
 
   /*******************************************************
-   * FUNCTOINS
+   * FUNCTIONS
    *******************************************************/
   makeDragAndDrop() {
     this.makeSortableNode(
