+++ course_flow/static/course_flow/js/react/src/Components/components/CommonComponents/EditableComponentWithComments.js
@@ -13,6 +13,7 @@ import { reloadCommentsAction } from '../../../redux/Reducers.js'
 // @components
 import Component from './Component.js'
 import ActionButton from './ActionButton.js'
+import EditableComponent from './EditableComponent.js'
 
 /*******************************************************
  * @CommentBox
