+++ course_flow/static/course_flow/js/react/src/Components/components/Strategy.js
@@ -4,7 +4,6 @@ import * as Constants from '../../Constants.js'
 import { Component } from './CommonComponents'
 import { getStrategyByID } from '../../redux/FindState.js'
 
-
 //Basic component to represent a Strategy
 class StrategyView extends Component {
   constructor(props) {
